2.2.Writing a Unit Test with Jest
The Power of Vue Test Utils
Now that we’ve scaffolded our tests, and got them both passing, we can switch out that “sanity test” with the real logic that can perform our given tests:

1.If a user is not logged in, do not show the logout button
2.If a user is logged in, show the logout button
In order to do this, we need the AppHeader component to be mounted (to check if the button is visible in the DOM or not). This would be quite a process to perform all on our own, but fortunately with the help of Vue Test Utils, it’s very simple because this library comes packaged with mount.
So let’s import mount into our testing file and see what it can do for us.
Above, in each of our tests we’ve created a wrapper const in which we mount(AppHeader). The reason we’re calling it wrapper is because in addition to mounting the component, this method creates a wrapper(https://vue-test-utils.vuejs.org/api/wrapper/) that includes methods to test the component. Of course, it’s helpful to understand the different properties and methods on the wrapper, so take some time to explore the documentation(https://vue-test-utils.vuejs.org/api/wrapper/).
Sidenote: In the Vue Test Utils you will also find the method shallowMount(). If your component has children, shallowMount() will return a simple implementation of that component instead of a fully rendered version. This is important because the focus of a unit test is the component in isolation and not the children of that component.
We’re still not performing our actual tests yet, but now we have a wrapper of the mounted AppHeader component, which we can use to write out the complete tests.
In our first test case, we know that by default the user is not logged in (our input is loggedIn: false) so we want to check and make sure the logout button is not visible.
To make assertions on the state of the logout button, we’ll need to get a reference to the button element that is defined in the template. To accomplish this we will rely on two methods available to us on our new wrapper: find() and isVisible(). The find() method will search through our template for a matching selector in order to locate our button, and isVisible() will return a boolean, telling us if that button is visible in our component or not.