2.3.Writing a Unit Test with Jest
Testing the Button’s visibility
Because we are testing the components behavior when we have a user (when loggedIn is true), we need to update that value or else this test will fail. How we do we this? Vue Test Utils to the rescue!
Here, we’re using the wrapper’s built-in setData() method to set our data to fit the correct scenario that we’re testing. However, if we were to run this test, it would fail.