3.1.Testing Props & User Interaction
In the previous lesson, we learned how to write and run our first unit tests. In this lesson, we’ll continue writing simple tests for a component that requires user interaction and takes in some props.

Random Number Component
In lesson 1, we identified the inputs and outputs of a component that generates a random number within the range of its min and max props.

What tests should we write?
In terms of this component’s inputs, it’s pretty obvious that the props are inputs since props are literally fed into the component. Another input would be the user interaction, whether a user clicked the button or not, which runs the method that generates the random number. The output is the rendered HTML displaying the randomNumber.

Inputs

Props:

  min & max
User Interaction:

  Clicking of the Generate Random Number button
Outputs

  Rendered Output (DOM)

Is the number displayed on the screen between min and max?
We can use this knowledge to figure out what to test in this component:

1.By default, the randomNumber data value should be 0
2.If we click the generate button, randomNumber should be between 1 (min) and 10 (max)
3.If we change the min and max props to 200 and 300 and click the button, randomNumber should be between 200 (min) and 300 (max).

Random Number Tests
To get started testing this component, we’ll created a new test file: /tests/unit/RandomNumber.spec.js
For now, we’ll simply scaffold out the tests and write default assertions that we know will fail. In the last lesson, we looked at scaffolding our tests with an assertion we know would pass. By using an assertion we know will definitely fail, this serves a similar purpose of making sure our components are behaving how we expect them to at first. Then, we’ll do the work to make the tests pass.
If we run npm run test:unit, we’ll see 3 failures for this test.