3.2.Testing Props & User Interaction
Checking the default random number
Looking at the component, we know that the default value for randomNumber is 0, so why would you even test for this? Well, what if someone else on our team changed the default randomNumber? Testing it gives us confidence 0 will always be displayed when the component is first loaded.
Our first step to test this is to mount the component we are testing (RandomNumber.vue), which gives us the wrapper that lets us dive into the component and test what we need to.
Here, we’re using the wrapper to get the html of our RandomNumber component, and asserting that we expect that html toContain a span with a 0 in its inner HTML.
If we run this test by typing npm run test:unit in our terminal, we’ll see that now only 2 tests are failing, and we’ve gotten our first test to pass. Now we can move on to our next test, which requires some user interaction.