3.3.Testing Props & User Interaction
Simulating User Interaction
We need to verify that when we click the generate random number button, we get a random number between the min and max props. The defaults for min and max are 1 and 10 respectively, so the random number should fall in that range.
Just as we have before, we’ll need to mount the RandomNumber component. The new concept here is that we need to trigger a button click on our generate random number button (which runs the method that uses the min and max props to generate a new randomNumber).
We’ll get a reference to the button element using the find() method, and then use the trigger() method to trigger an event on the wrapper DOM node. The first and required argument to the trigger method is a string for what event type to trigger. In this case, we want to trigger a 'click' event on the button.
Now that the button was ‘clicked’, it should have generated a random number.
This effectively finds the span, and accesses that element’s text content. But because we need that content to be an integer, we’ve wrapped in in parseInt.
Finally, we can use some Jest assertions to make sure the random number falls between the min prop 1 and max prop 10
If we run the tests now, we should see 2 passing. Now we can work on the final test.