3.4.Testing Props & User Interaction
Setting different prop values
Since someone using this component can change the min and max value range via the min and max props, we need to test for this. To do this, we’ll use the mount() method, which takes an optional second argument, where we can pass in some options, including propsData. This can be used to reset the min and max values, to 200 and 300 respectively, in our case.
With our new min and max, the test is going to look pretty similar to the one we just wrote above.
At this point you can run your tests and all of them should pass.

Let’s Revue
Not all components are created equally, so testing them often means we’ll have to consider things like simulating button clicks, and testing props. In our next lesson,we’ll continue learning about testing common aspects of Vue components.