4.1. Testing Emitted Events
In the previous lesson, we looked at testing a component that took in some props and generated a number based upon a button click. This required us to simulate (trigger) the button click within our test. That “click” falls under the category of a native DOM event, but often in Vue components, we need our components to emit their own custom event, and in this lesson we’ll look at testing such events.

What’s a custom event?
If custom events are new to you, you can check out the Communicating Events lesson in our Intro to Vue.js course. In short: sometimes a child component needs to let another component in the app know that something happened within it. It can broadcast that something happened by emitting a custom event (such as formSubmitted) to let its parent know that event happened. The parent can wait, listening for that event to happen, then respond accordingly when it does.

The Starting Code
Fortunately, Vue Test Utils provides us with an emitted API, which we can use to test these kinds of custom emitted events within our components. We’ll explore how to use that API in a moment, but first let’s look at the component that we’ll be testing.
As you can see, we a very simple login form. Pay attention to how we have the onSubmit() method, which is used to $emit a custom event by the name of formSubmitted, which sends a payload containing the name data, which is bound to our input element. We want to test that when the form iis submitted, it indeed emits an event, containing the user name payload.

Scaffolding the test file
When writing tests, it’s helpful if we write them in a way that mimics how an actual end user will interact with the component. So how will a user use this component? Well, they will find the text input field, then they’ll add their name, then submit the form. So we’ll replicate that process as closely as possible in our test.
Above, you can see that we’ve imported the LoginFrom component along with mount from Vue Test Utils, and we need to accomplish the following steps:

Find text input
1.Set value for text input
2.Simulate form submission
3.Assert event has been emitted
4.Assert payload is correct
Let’s implement these steps.