4.2. Testing Emitted Events
Setting the text input value
First, just like our end user would, we’ll find the text input and set its value.

A note about targeting inputs
This works great for our specific needs, but it’s worth mentioning here that in production tests, you might considering using test-specific attribute on your elements, like so:

<input data-testid="name-input" type=text v-model="name" />
Then, in your test file, you’d find the input using that attribute.

const input = wrapper.find('[data-testid="name-input"]')
This is beneficial for a few reasons. First, if we had multiple inputs we could target them specifically with these ids, and perhaps more importantly this decouples the DOM from your test. For example, if you eventually replaced the native input with an input from a component library, the test would still conform to the same interface and wouldn’t need to change. It also solves the issue of a designer changing the class or id name of the element, causing your test to fail. Test-specific attributes are one way to future-proof your tests.

Simulating the form submission
Once our end user has filled out our form, the next step they would take would be to submit the form. In the previous lesson, we looked at how the trigger method can be used to simulate an event on a DOM element.
While you might be tempted to use trigger on our form’s button to simulate our form submission, there’s a potential problem with that. What if we eventually removed the button from this component and instead relied on the input’s keyup.enter event to submit the form? We would have to refactor our test. In other words: in that case our test would have been too tightly coupled with the implementation details of our component’s form. So the more future-proofed solution would be to force a submit even on the form itself, without relying on our button as the middle man.
Now by using wrapper.trigger('submit'), we’ve implemented a more scalable, decoupled solution to simulate a user submitting our form.

Testing our expectations
Now that our input field’s value has been set and our form has been submitted, we can move on to testing that what we expect to happen is actually happening:

  The event has been emitted
  The payload is correct
Here we are using Vue Test Utils’ emitted API(https://vue-test-utils.vuejs.org/api/wrapper/emitted.html) to store any calls of the formSubmitted event in a const, and asserting that we expect that array to have a length of 1. In other words: we are checking to see if the event was indeed emitted.
Now we just need to confirm that the event was emitted with the proper payload (our component’s name data value). We’ll use the emitted API again for this.

If we were to console.log wrapper.emitted('formSubmitted'), we’d see this:

[ [ { name: 'Adam Jahr' } ] ]
So in order to target the payload itself, the syntax would look like:

wrapper.emitted('formSubmitted')[0][0])
We’ll then match that against our expected payload, which for organization purposes we’ll store in const expectedPayload = { name: 'Adam Jahr' }

Now we can check to see that our expectedPayload object matches the payload emitted along with the formSubmitted event.
Sidenote: We could alternatively hard-code the expected payload into the matcher: .toEqual({ name: 'Adam Jahr' }). But storing it in a const keeps us a bit more clear about what is what.
If we npm run test:unit we’ll see that our new test is passing. Great work!

Let’s ReVue
We’ve learned how to write a test while replicating how a user would interact with a component, in order to test that our custom events are emitted with the correct payload.