5.1.Testing API Calls Unless you’re working with a simple static website, your Vue app is likely making API calls from within certain components. In this lesson, we’ll look at how we can test these kinds of data-fetching components. The first thing to understand about testing components that make API calls is that we don’t want to be making real calls out to our backend. Doing so would couple our unit tests to the backend. This becomes an issue when we want to execute our unit tests in Continuous Integration(https://en.wikipedia.org/wiki/Continuous_integration). Real backends can also be unreliable, and we need our tests to behave predictably. We want our tests to be fast and reliable, and we can achieve that by mocking our API calls and simply focusing on the inputs and outputs of the component we’re testing. In this lesson we’ll be using axios, the popular promise-based HTTP client, to make our calls. This means we’ll have to mock axios’ behavior, which we’ll do in a moment. But first let’s take a look at the starting code.
The Starting Code
For the sake of simplicity, instead of plugging into a full backend, we’re using json-server, which gives us a fake REST API. If this library is new to you, we taught it over in our Real World Vue course. What you need to know for this lesson is: our db.json file is our database, and json-server can fetch data from it.
Our simple db has one endpoint: “message”, and this is the data we’ll be fetching.
In our project, I’ve also added an API service layer, which will handle the actual API calls.
As you can see, we’ve imported axios and we are exporting the getMessage() function, which makes a get request to our endpoint: http://localhost:3000/message, then we’re returning the data from the response.
Now that we understand how and where we’re pulling our data from, we can look at the component that triggers this API call, and displays the returned data.
In the <script> section, we’ve imported the getMessage function from our axios.js file, and when our component is created, it calls getMessage using async / await since axios is asynchronous and we need to wait for the promise it returns to resolve. When it resolves, we’re setting our component’s local message data equal to the resolved value, which gets displayed in the template.
We’re also wrapping the getMessage call with try in order to catch errors that might happen, and if one does occur, we’re setting our local error data accordingly and displaying that error.

Inputs & Outputs
Looking at the MessageDisplay.vue component, what are the inputs and outputs that we’ll need to consider when writing our test?

Well, we know that the response from the getMessage call is our input, and we have two possible outputs:

1.The call happens successfully and the message is displayed
2.The call fails and the error is displayed

So in our test file, we’ll need to:

1.Mock a successful call to getMessage, checking that the message is displayed

2.Mock a failed call to getMessage, checking that the error is displayed

Let’s get started with learning how to mock axios.