5.3.Testing API Calls -- npm i flush-promises axios --save-dev
Awaiting Promises
When figuring out where to await in our test, we have to think back to how getMessage is being called in the component we’re testing. Remember, it’s being called on the component’s created lifecycle hook?

📄MessageDisplay.vue

async created() {
            try {
               this.message = await getMessage()
            } catch (err) {
               this.error = 'Oops! Something went wrong.'
            }
       }
Since vue-test-utils doesn’t have access to the internals of promises that are enqueued by the created lifecycle hook, we can’t really tap into anything to await for that promise. So the solution here is to use a third-party library called flush-promises which allows us to—well—flush the promises, ensuring they’re all resolved prior to running our assertions.