5.4.Testing API Calls
Once we’ve installed the library with npm i flush-promises --save-dev, we’ll import it into our testing file and await the flushing of the promises.
Now that we’ve ensured promises will be resolved before our assertions are run, we can write those assertions.

Our Assertions
First up, we’ll want to make sure our API call is only happening once.
We’re simply running the method .toHaveBeenCalledTimes() and passing in the number of times we expect getMessage to have been called: 1. Now we’ve ensured that we aren’t accidentally hitting our server more times than we should be.
Next up, we need to check that our component is displaying the message it received from our getMessage request. In the MessageDisplay component’s template, the p tag that displays the message has an id to be used for tests: data-testid="message"
We learned about these testing ids in the previous lesson. We’ll use that id to find the element, then assert that its text content should be equal to the value our mocked getMessage request resolved with: mockMessage
If we run npm run test:unit in the terminal, we’ll see our newly written test is passing! We can now move on to our second test, where we’ll mock a failed getMessage request and check that our component is displaying the error.