5.5.Testing API Calls
Mocking a failed request
The first step, of mocking the failed API call, is very similar to our first test.
Notice how we’re using mockRejectedValueOnce to simulate the failed get request, and we’re passing it the mockError for it to resolve with.
After awaiting the flushing of the promises, we can then check that the call only happened once and verify that our component’s template is displaying the expected mockError.
Just like our first test, we’re using .toHaveBeenCalledTimes(1) to make sure we’re not making the API call more than we should be, and we’re finding the element that displays the error message and checking its text content against the mockError that our mocked failed request returned.

Now if we run these tests, what happens? The test is failing:

Expected number of calls: 1 Received number of calls: 2

Hmm… what’s happening here? Well, in our first test, getMessage was called, and then it gets called again in our second test. We haven’t done anything to clear out our mocked getMessage function before running the second test.